Rename files aux and and replace it like dependency in files with aux dependency to solve problem cloning repository in Windows
