+++ templates/dbs_connections/s3_plot.py
@@ -2,7 +2,7 @@ from os import getenv, environ
 
 import boto3
 import pandas as pd
-from aux import init_sdk
+from app_shimoku import init_sdk
 
 environ["AWS_ACCESS_KEY_ID"] = getenv('AWS_ACCESS_KEY_ID')
 environ["AWS_SECRET_ACCESS_KEY"] = getenv('AWS_SECRET_ACCESS_KEY')
